feat: add some user validations and  appError to handle async errors

+++ src/modules/users/useCases/deleteUser/deleteUserUseCase.ts
@@ -1,3 +1,4 @@
+import { AppError } from '@errors/appError';
 import { UsersRepository } from '@modules/users/repositories/implementation/usersRepository';
 import { inject, injectable } from 'tsyringe'
 
@@ -8,6 +9,10 @@ export class DeleteUserUseCase {
         private usersRepository: UsersRepository
     ) { }
     async execute(id: string) {
+        const user = await this.usersRepository.findById(id)
+        if (!user) {
+            throw new AppError(404, "Usuário não encontrado.")
+        }
         await this.usersRepository.deleteUser(id)
     }
 }
